chore: consolidate tooling and polish robotic bees
