+++ .github/workflows/benchmarks.yml
@@ -85,7 +85,7 @@ jobs:
       - name: Run Lighthouse CI
         uses: treosh/lighthouse-ci-action@3e7e23fb74242897f95c0ba9cabad3d0227b9b18 # 12.6.2
         with:
-          configPath: ./lighthouserc.json
+          configPath: ./tools/configs/lighthouserc.json
           uploadArtifacts: false
           temporaryPublicStorage: false
       - name: Upload Lighthouse artifacts
