+++ biome.json
@@ -1,28 +1,4 @@
 {
   "$schema": "https://biomejs.dev/schemas/2.5.2/schema.json",
-  "files": {
-    "includes": [
-      "src/**",
-      "tests/**",
-      "scripts/**",
-      "*.ts",
-      "*.tsx",
-      "*.json",
-      "!src/features/legacy-three/legacy-main.ts",
-      "!src/shared/styles/legacy.css"
-    ]
-  },
-  "formatter": {
-    "enabled": true,
-    "indentStyle": "space",
-    "indentWidth": 2
-  },
-  "linter": {
-    "enabled": true
-  },
-  "css": {
-    "parser": {
-      "tailwindDirectives": true
-    }
-  }
+  "extends": ["./tools/configs/biome.json"]
 }
