+++ package.json
@@ -5,36 +5,36 @@
   "type": "module",
   "packageManager": "npm@11.12.1",
   "scripts": {
-    "dev": "rsbuild dev",
+    "dev": "rsbuild dev --config tools/configs/rsbuild.config.ts",
     "prebuild": "npm run check:no-js && npm run validate:content && npm run seo:check",
-    "build": "rsbuild build",
-    "preview": "rsbuild preview",
+    "build": "rsbuild build --config tools/configs/rsbuild.config.ts",
+    "preview": "rsbuild preview --config tools/configs/rsbuild.config.ts",
     "ci": "npm run ci:quality && npm run ci:build",
     "ci:quality": "npm run check:no-js && npm run validate:content && npm run seo:check && npm run typecheck && npm run lint && npm run format:check && npm run test:unit",
-    "ci:build": "npm run build && npm run check:budgets && node scripts/collect-benchmarks.ts --out artifacts/benchmarks/build.json",
+    "ci:build": "npm run build && npm run check:budgets && node tools/scripts/collect-benchmarks.ts --out artifacts/benchmarks/build.json",
     "ci:security": "npm run check:workflows && npm run security:audit",
     "typecheck": "tsc --noEmit",
     "lint": "biome check .",
     "lint:fix": "biome check --write .",
     "format": "biome format --write .",
     "format:check": "biome format .",
-    "check:no-js": "node scripts/assert-no-js.ts",
-    "check:actions-pinned": "node scripts/assert-actions-pinned.ts",
+    "check:no-js": "node tools/scripts/assert-no-js.ts",
+    "check:actions-pinned": "node tools/scripts/assert-actions-pinned.ts",
     "check:workflows": "npm run check:actions-pinned",
-    "check:budgets": "node scripts/collect-benchmarks.ts --check-budgets",
-    "validate:content": "node scripts/validate-content.ts",
-    "seo:generate": "node scripts/generate-seo.ts",
-    "seo:validate": "node scripts/validate-seo.ts",
+    "check:budgets": "node tools/scripts/collect-benchmarks.ts --check-budgets",
+    "validate:content": "node tools/scripts/validate-content.ts",
+    "seo:generate": "node tools/scripts/generate-seo.ts",
+    "seo:validate": "node tools/scripts/validate-seo.ts",
     "seo:check": "npm run seo:validate",
-    "test": "vitest run",
-    "test:unit": "vitest run",
-    "test:component": "cypress run --component",
-    "test:e2e": "cypress run --e2e",
+    "test": "vitest run --config tools/configs/vitest.config.ts",
+    "test:unit": "vitest run --config tools/configs/vitest.config.ts",
+    "test:component": "cypress run --project . --config-file tools/configs/cypress.config.ts --component",
+    "test:e2e": "cypress run --project . --config-file tools/configs/cypress.config.ts --e2e",
     "test:ci": "npm run ci:quality && npm run ci:build",
     "analyze": "npm run benchmark:build",
     "benchmark": "npm run benchmark:build",
-    "benchmark:build": "npm run build && node scripts/collect-benchmarks.ts --out artifacts/benchmarks/build.json",
-    "assets:optimize": "node scripts/optimize-assets.ts",
+    "benchmark:build": "npm run build && node tools/scripts/collect-benchmarks.ts --out artifacts/benchmarks/build.json",
+    "assets:optimize": "node tools/scripts/optimize-assets.ts",
     "security:audit": "npm audit --audit-level=moderate"
   },
   "dependencies": {
