+++ src/shared/styles/tokens.css
@@ -8,8 +8,8 @@
 :root {
   --site-scale: 1;
   --nav-scale: var(--site-scale);
-  --join-us-size: clamp(1rem, 0.75rem + 1vw, 1.45rem);
-  --bee-line-width: clamp(13.5rem, 18vmin, 16.5rem);
-  --orb-size: clamp(0.62rem, 0.44rem + 0.58vw, 0.96rem);
-  --social-box-size: clamp(7rem, 18vmin, 13rem);
+  --join-us-size: clamp(1.15rem, 0.85rem + 1.15vw, 1.75rem);
+  --bee-line-width: clamp(16rem, 22vmin, 21rem);
+  --orb-size: clamp(0.78rem, 0.52rem + 0.75vw, 1.18rem);
+  --social-box-size: clamp(8rem, 20vmin, 14.5rem);
 }
