+++ tools/configs/biome.json
@@ -1,0 +1,30 @@
+{
+  "$schema": "https://biomejs.dev/schemas/2.5.2/schema.json",
+  "root": false,
+  "files": {
+    "includes": [
+      "src/**",
+      "tests/**",
+      "tools/**",
+      "cypress/**",
+      "*.ts",
+      "*.tsx",
+      "*.json",
+      "!src/features/legacy-three/legacy-main.ts",
+      "!src/shared/styles/legacy.css"
+    ]
+  },
+  "formatter": {
+    "enabled": true,
+    "indentStyle": "space",
+    "indentWidth": 2
+  },
+  "linter": {
+    "enabled": true
+  },
+  "css": {
+    "parser": {
+      "tailwindDirectives": true
+    }
+  }
+}
