+++ tools/configs/cypress-vite.config.ts
@@ -1,0 +1,17 @@
+import { fileURLToPath } from "node:url";
+
+export default {
+  configFile: false,
+  publicDir: fileURLToPath(new URL("../../public", import.meta.url)),
+  resolve: {
+    alias: {
+      "@": fileURLToPath(new URL("../../src", import.meta.url)),
+    },
+  },
+  esbuild: {
+    jsx: "automatic",
+  },
+  define: {
+    "process.env.NODE_ENV": JSON.stringify("test"),
+  },
+};
