+++ tools/scripts/assert-actions-pinned.ts
