+++ tools/scripts/assert-no-js.ts
