+++ tools/scripts/optimize-assets.ts
