+++ tsconfig.json
@@ -19,12 +19,5 @@
     "jsx": "react-jsx",
     "types": ["node", "vitest/globals", "cypress"]
   },
-  "include": [
-    "src",
-    "scripts",
-    "tests",
-    "rsbuild.config.ts",
-    "vitest.config.ts",
-    "cypress.config.ts"
-  ]
+  "include": ["src", "tests", "tools"]
 }
